requirements.txt
pandas==0.24.2
Flask==1.1.1
matplotlib==3.1.0
numpy==1.18.1
boto3==1.13.3
seaborn==0.9.0
pytest==5.0.1
Flask_SQLAlchemy==2.4.3
PyYAML==5.3.1
scikit_learn==0.23.1
sphinx_rtd_theme==0.4.3
PyMySQL==0.9.3
